requirements.txt
playwright==1.35.0
requests==2.31.0
openai==0.27.0
pyyaml==6.0
loguru==0.7.0
pytest==7.4.0
